Remove redundant cast from SmallintType and TinyintType

+++ core/trino-spi/src/main/java/io/trino/spi/type/TinyintType.java
@@ -138,7 +138,7 @@ public final class TinyintType
     @Override
     public long getLong(Block block, int position)
     {
-        return (long) block.getByte(position, 0);
+        return block.getByte(position, 0);
     }
 
     @Override
